chore: create placements consts
This allows for better usage of the types and enums across all components that need it
